requirements.txt
gitpython==3.1.27
pandas==1.4.1